# Conflicts:
#	src/CargoBody.h
#	src/HyperspaceCloud.h
#	src/Planet.h
#	src/Projectile.h
#	src/Ship.h
#	src/ShipCockpit.h
#	src/SpaceStation.h
#	src/Star.h
#	src/TerrainBody.h
#	src/scenegraph/Model.h
#	win32/vs2015/pioneer.vcxproj.filters

+++ src/Frame.h
@@ -1,11 +1,10 @@
-// Copyright © 2008-2016 Pioneer Developers. See AUTHORS.txt for details
+// Copyright © 2008-2018 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #ifndef _FRAME_H
 #define _FRAME_H
 
 #include "libs.h"
-#include "Serializer.h"
 #include "IterationProxy.h"
 #include "json/json.h"
 #include <string>
@@ -62,7 +61,7 @@ public:
 	void AddChild(Frame *f) { m_children.push_back(f); }
 	void RemoveChild(Frame *f);
 	bool HasChildren() const { return !m_children.empty(); }
-	unsigned GetNumChildren() const { return m_children.size(); }
+	unsigned GetNumChildren() const { return static_cast<Uint32>(m_children.size()); }
 	IterationProxy<std::vector<Frame*> > GetChildren() { return MakeIterationProxy(m_children); }
 	const IterationProxy<const std::vector<Frame*> > GetChildren() const { return MakeIterationProxy(m_children); }
